Writing metrics and reading config from consul.

+++ spectator/Sources/IQueryableSource.cs
@@ -2,6 +2,6 @@
 {
     public interface IQueryableSource
     {
-        int QueryValue(string path);
+        double QueryValue(string path);
     }
 }
